Move LocalDirectoryDependency to another package
