+++ src/test/groovy/com/github/blindpirate/gogradle/core/dependency/parse/DirMapNotationParserTest.groovy
@@ -2,7 +2,7 @@ package com.github.blindpirate.gogradle.core.dependency.parse
 
 import com.github.blindpirate.gogradle.GogradleRunner
 import com.github.blindpirate.gogradle.core.dependency.GolangDependency
-import com.github.blindpirate.gogradle.core.pack.LocalDirectoryDependency
+import com.github.blindpirate.gogradle.core.dependency.LocalDirectoryDependency
 import com.github.blindpirate.gogradle.support.WithResource
 import org.junit.Test
 import org.junit.runner.RunWith
